Sequental port selection in process-based fl runner
